pyproject.toml
[tool.poetry]
name = "python-rconfig"
version = "20.1.1"
description = "Helps get configuration from consul server"
license = "MIT"
authors = ["Artem Angelchev <artangelchev@gmail.com>"]
homepage = "https://github.com/ArtemAngelchev/python-rconfig"
repository = "https://github.com/ArtemAngelchev/python-rconfig"
packages = [{ include = "rconfig" }]
readme = "README.rst"
keywords = [
  "configurations",
  "settings",
  "consul",
  "env",
  "python",
  "environment-variables",
  "configuration-managment"
]
classifiers = [
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.6",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Intended Audience :: Developers",
  "Intended Audience :: System Administrators",
  "Operating System :: OS Independent",
  "Topic :: System :: Systems Administration",
  "Topic :: Utilities",
  "Environment :: Web Environment",
]

[tool.poetry.dependencies]
python = ">=3.6, <4"
python-consul = "^1.1.0"

click = {version = "^7.0", optional = true}

[tool.poetry.dev-dependencies]
pytest = "^5.3.5"
pytest-cov = "^2.8.1"
pylint = "^2.4.4"
flake8 = "^3.7.9"
sphinx = "^2.4.4"

[tool.poetry.extras]
cli = ["click"]

[tool.poetry.scripts]
rconfig = "rconfig.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
